Tests: Drivers: Timeout: Extracted test templates
Test cases will be reused in LowPowerTimeout tests.
